requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
django-sortedm2m==3.1.1
django-autoslug==1.9.9
django-jazzmin==2.6.1
djoser==2.2.2
django-debug-toolbar==4.2.0
django-filter==23.5
argon2-cffi==23.1.0
pytils==0.4.1
celery==5.3.6
django-redis==5.4.0
django-celery-results==2.5.1
gevent==23.9.1
django-celery-beat==2.1.0
flower==2.0.1
django-import-export==3.3.7